много много
-поддержка дабл полей
-исправлены UI баги
-маска валюты
-конвертирование валюты

+++ app/src/main/java/dizel/budget_control/budget/use_cases/CreateBudgetUseCase.kt
@@ -57,7 +57,7 @@ class CreateBudgetUseCase(
 
     data class Params(
         val title: String,
-        val money: Long,
+        val money: Double,
         val currency: Currency
     )
 }
